+++ src/services/book.service.test.ts
@@ -1,0 +1,7 @@
+import { BookService } from "./book.service";
+
+describe("BookService", () => {
+  it("BookService should be defined", () => {
+    expect(BookService).toBeDefined();
+  });
+});